+++ native/quanttrace/src/hello.c
@@ -9,7 +9,7 @@
 #  define QT_EXPORT __attribute__((visibility("default")))
 #endif
 
-static const char QT_VERSION[] = "0.0.2-cube-f12";
+static const char QT_VERSION[] = "0.0.3-depsgraph";
 
 QT_EXPORT const char *quanttrace_version(void)
 {